Added GL_EXT_nonuniform_qualifier to fragment shader because texture id was interpolated from another traingle so the texture id at the edges changed to the other id so it appeared as artifacts because often the blank texture was sampled

+++ Renderer/Source/RendCore.h
@@ -16,7 +16,7 @@
 #include <array>
 #include <set>
 
-#define VULKANDEBUG(x,y) x== VK_SUCCESS? (void)0: Core::Log(ErrorType::Error,y,": ",(int)x) 
+#define VULKANDEBUG(x,y) x== VK_SUCCESS? (void)0: Core::Log(ErrorType::Error,y,": ",(int64_t)x) 
 #define MAX_FRAME_DRAWS 3
